Convert pink noise generator from 'int' to 'int32_t'.
Add test cases for creating pink noise with 16/32 bits.  Fix a bug
where "(1<<32)-1" gets calculated differently on 32bit Intel vs 32bit
Arm, causing the test cases to fail on certain platforms.

+++ mythtv/libs/libmyth/audio/pink.h
@@ -5,15 +5,15 @@
 
 #define PINK_MAX_RANDOM_ROWS   (30)
 #define PINK_RANDOM_BITS       (24)
-#define PINK_RANDOM_SHIFT      ((sizeof(long)*8)-PINK_RANDOM_BITS)
+#define PINK_RANDOM_SHIFT      ((sizeof(int32_t)*8)-PINK_RANDOM_BITS)
 
 // NOLINTNEXTLINE(modernize-use-using)
 typedef struct
 {
-  long      pink_rows[PINK_MAX_RANDOM_ROWS];
-  long      pink_running_sum;   /* Used to optimize summing of generators. */
-  int       pink_index;        /* Incremented each sample. */
-  int       pink_index_mask;    /* Index wrapped by ANDing with this mask. */
+  int32_t   pink_rows[PINK_MAX_RANDOM_ROWS];
+  int32_t   pink_running_sum;  /* Used to optimize summing of generators. */
+  int32_t   pink_index;        /* Incremented each sample. */
+  int32_t   pink_index_mask;   /* Index wrapped by ANDing with this mask. */
   float     pink_scalar;       /* Used to scale within range of -1.0 to +1.0 */
 } pink_noise_t;
